+++ Changelog.md
@@ -25,9 +25,10 @@
 - Added the `ProductAvailabilityScope` enum class with `listing`, `viewing` and `buying` values
 - Added the `gtin` field to the products and master product variants tables
 - Fixed the missing float cast of the order item's price field
-- PHP 8.4 deprecation notice fixes
+- Fixed PHP 8.4 deprecation notices
 - Added support for calculating item-level promotions on checkout
 - Added ETA fields to Shipping Methods
+- Added a customizable Shipping Method eligibility checker class
 - Added the Order::inChannel() scope (Foundation)
 - Added the Video module
 - Added the Translation module